scraper/pyproject.toml
[tool.poetry]
name = "vensers_journal_mtr_scraper"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.10.0"
requests = "^2.27.1"
"pdfminer.six" = "^20211012"
pyparsing = "^3.0.6"
pytest = "^6.2.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mtr-scrape = 'vensers_journal_mtr_scraper.mtr:main'